♻️ [Refactoring]: figma-node.test.tsを機能別に3つのファイルに分割
- figma-node-type-guards.test.ts: 型ガードのテスト
- figma-node-factory-methods.test.ts: ファクトリーメソッドのテスト
- figma-node-style-helpers.test.ts: スタイルヘルパーのテスト

describeを使用しない形式に統一
